hailo_requirements.txt
# hailort and tappas through install script
torch
librosa
onnxruntime
faster_whisper
transformers
jiwer
psutil
pyaudio
silero_vad
useful-moonshine-onnx@git+https://git@github.com/usefulsensors/moonshine.git#subdirectory=moonshine-onnx